requirements.txt
asgiref==3.7.2
Django==5.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
gunicorn==21.2.0
whitenoise==6.6.0